relative imports replaced with normal ones. `print_greeting_msg` now in separate file. `enable_trace_points` call (wrong) replaced with empty_file loading (ugly but should always work). Including gems into $LOAD_PATH.
